feat: CandleStickIndentifier 타입 생성 및 CandleStickData 타입을 별도 파일로 분리
